Additional statement needed for reading logs
In order to read logs one additional statement is needed,
`allow group grafana to read compartments in tenancy`

As it has been tested that adding this one will allow grafana-user to read the data. 

It can be tested in OCI console under Observability & Management > Logging > Search.

By adding only the 3 statements documented grafana-user will not be allowed to search any logs, after adding the fourth statement it works as expected.